// src/types/generated/SegmentState.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AgentType } from "./AgentType";

export type SegmentState = { agent_id: number, segment_id: number, parent_segment_id: number | null, x: number, y: number, z: number, yaw: number, pitch: number, roll: number, joint_anchor_x: number, joint_anchor_y: number, joint_anchor_z: number, joint_axis_x: number, joint_axis_y: number, joint_axis_z: number, energy: number, hydration: number, head_direction: [number, number, number], agent_type: AgentType | null, };